[dart2wasm] Avoid array allocation when ending json object
When the json parser sees the end of a json object (i.e. the '}') it
currently creates a growable array in `popContainer()` but at the call
site it doesn't need the growable array, so it just gets the wasm array
and the length out.

We can avoid this array allocation by just directly accessing the
current container wasm array & length.

Change-Id: Ib66a3ddf98934c3c083347ff41c93ed4de2eeaf5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/409960
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>