[dart2wasm] Avoid indirect call to `key.hashCode` in json map creations
After a recent change we started to eagerly calculate the hash code of
json object keys so when we create the string objects we initialize the
hash as well (as we know it will be needed to create the map).

This means that the actual creation of the hash map can do less work.
Right now it does an indirect call to `key.hashCode`, but since we know
those particular maps from the json decoder come only with string keys
and most of them will have their hash code already prepopulated we can
get avoid the indirect call to `key.hashCode` in most cases.

Change-Id: I38987281bd8b216832cac950ea74d0ca201ff651
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/409980
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>