[dart2wasm] Make json parser use JS `parseFloat()` directly instead of `double.parse()`
The `double.parse()` function first runs a regular expression over the
string to validate it's a valid string representation of a double.

Though the json parser does this kind of validation itself, so we can
avoid this regexp match and call the JavaScript `parseFloat()` function
directly.

Change-Id: Ib2c2c4789d63b75b9a24eac5b5ee2b3483c65327
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/410320
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>