[stable][cfe] Issue 60121: Fix bug with isWildcard flag for formals.
Issue description: Users are finding that their debug variables are
evaluated to the wrong values and it's due to wildcard variables being
included, but being unable to be evaluated. Wildcard variables shouldn't
be able to be evaluated in the first place.

What is the fix: Correctly propagate the `isWildcard` flag in the CFE to
formal parameters. The VM uses this flag to determine if the variable
can be excluded from debugging evaluation. The flag is now set correctly
with the fix and won't appear in VM debug.

Why cherry-pick: This is causing debug variables to be evaluated
incorrectly because wildcard variables are appearing in the debug panel
of variables and shouldn't be.

Risk: Low, fixes wildcard variables that are formal parameters and touches no other code. This is code that should've been in the feature from the start.

Issue link(s): https://github.com/dart-lang/sdk/issues/60121

Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/410161
Change-Id: Ie7aa3561183e6e326d65c09165a132d6bb3e72a2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/410640
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Siva Annamalai <asiva@google.com>
Commit-Queue: Kevin Chisholm <kevinjchisholm@google.com>