Update date_time_extremes_test to not make assumptions.
The test assumes that seconds, milliseconds and occasionally
microseconds will necessarily be zero, which is to optimistic
in practice.

The changed test will use all the values from a source
`DateTime`, and not stop when it thinks it knows the rest.
(Even for UTC values, where it actually does know.)

Change-Id: Ie88e6c33e721f9c7cfba605f0e2cb5f4d2f72218
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/411021
Reviewed-by: Nicholas Shahan <nshahan@google.com>
Commit-Queue: Lasse Nielsen <lrn@google.com>