Fine. Read summary bundle bytes if its requirements are satisfied.
Similarly, read diagnostics from bytes if requirements are still
satisfied.

Note, not all requirements are recorded yet, so if enabled now, we might get stale diagnostics. But I believe that with time the implementation
can be worked to the point when we record all necessary requirements.

Change-Id: Ie81a1c4aa5d41cba3e181e85a30970f2beb97f58
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/405820
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>