Elements. Avoid List.cast() mostly when create InterfaceTypeImpl.
This is a bit tricky because the API for `instantiate()` is that we
accept `List<DartType> typeArguments`, and the clients generally will
give us `List<DartType>`, not `List<TypeImpl>`. But we want `List<TypeImpl>`.

So, I added `instantiateImpl`, and use it in the analyzer, and avoid a forcing `<TypeImpl>[]`, type check and `cast<TypeImpl>()`.

Change-Id: I5e8e0ed77853b88e8a1dbc6360ab8ad29b803206
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/412782
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>