[vm] Make FFI implementation less fragile wrt field order
Replace access of certain fields by hardcoded indices with
lookup by name. As number of fields in those classes is small,
the lookup should not take considerable time.

This fix is done in anticipation for the CFE change
https://dart-review.googlesource.com/c/sdk/+/412880
which may change order of declared fields if there are patches.

TEST=manually tested with CFE change

Change-Id: Ibf4ab6719ee263711fb7c6b21b038ea1dc5f9e7b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/412803
Reviewed-by: Daco Harkes <dacoharkes@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>