[analyzer] Fix dead code reporting in property access and is/as.
The way dead code detection works in the analyzer is that at various
ad hoc times during the process of resolution (typically before
visiting an expression or statement), the `ResolverVisitor` calls
`checkUnreachableNode`, which asks flow analysis whether the current
point in execution is reachable, and if it isn't, arranges for an
unreachable code warning to be produced.

This wasn't working for property extractions, type casts, and type
tests because there's no subexpression whose execution follows the
subexpression that returns `Never`, so `checkUnreachableNode` wasn't
getting called.

The fix is simple; just add some calls to `checkUnreachableNode` to
cover the RHS of a property access, type cast, or type test.

Note that property accesses in the analyzer can be represented either
as `PrefixedIdentifier` or `PropertyAccess`, so the property access
fix had to be done in two places (and accordingly there are two tests
for it).

Fixes https://github.com/dart-lang/sdk/issues/60247.

Bug: https://github.com/dart-lang/sdk/issues/60247
Change-Id: I49e739784f629b77443ce6d31e67809f48de75be
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/413521
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>