[infra] Stop specifying a particular version of Ubuntu for Linux builders by default.
Test shards already have this weaker constraint. Weakening this constraint makes it easier to upgrade machines to the next Ubunutu LTS release, which would otherwise require changes in this repository and the GCE configuration to be deployed at the same time or a staged upgrade that temporarily reduces the effective pool capacity. Since Dart is supposed to support all the LTS releases in the machine pool anyway, it seems appropriate for the main shards to also have the weaker constraint.

Change-Id: If2744d36f838f00f238f0f9d38074a27f7fa5307
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/413323
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Ryan Macnak <rmacnak@google.com>