[analyzer] Add Fragment.offset getter.
This getter behaves similarly to `nameOffset`, however, if the
fragment is synthetic or does not have a name, it returns a suitably
chosen canonical offset. The intention is to provide a convenient API
that clients can use to point the user to the any fragment, saving
them from having to write special case code to handle the less common
cases of synthetic and unnamed fragments.

Change-Id: I42187a47053904ff73700698a670debe91082938
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/411660
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>