[analysis server] Sort members more consistant
I ran into an issue (in a child CL) where I couldn't satisfy the sort
test because it kept changing the ordering of "p" and "P" because the
sorting does `.toLowerCase` before comparing.
In this CL - when the result of comparing with a `.toLowerCase` is 0 it
compares without the `.toLowerCase` call.

Change-Id: Ieed0f199cb61ea45d475c5271a127fbf4463edb7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/413420
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>