[gardening] Make standalone/regress_52715_test more robust
The `vm-service` is starting somewhat concurrently with the `main` isolate
and the `main` can therefore not assume it's the only one allocating
memory in the process.

By delaying the measurement logic in `main`, it's more likely the
`vm-service` will have done it's work and reached steady memory state.

This makes the test more robust.

Change-Id: I5a641ad98ba269b8c5961ea6d20b06223621a4e2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/413641
Reviewed-by: Daco Harkes <dacoharkes@google.com>