[dart2wasm] Optimize fused json+utf8 decoder performance
Somewhat similar to what we already do for json object keys, we do now
for values: When json decoding from bytes, we look at all the bytes from
the string already (e.g. to find out if it's ASCII) so we may as well
remember those bytes in a WasmArray<WasmI16> (which we need later on
when creating the JS string). Doing so will avoid another O(n) pass over
the bytes.

We don't do this when json decoding from string, because here we can
use JS substring API (i.e. we don't have to create a JS string from a
WasmArray<WasmI16>).

Due to the "messiness" of this CL, I've pulled it out of the CL
that enabled JS strings by-default into this (separate) CL.

=> This brings back performance for utf8+json fused decoders
=> Significant improvements in `JsonDecode.*FromBytes` benchmarks

Change-Id: I8f6b8b26812bd789b06c82ad4a7404855bec10bd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/412642
Reviewed-by: Lasse Nielsen <lrn@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>