Remove pkg/OWNERS.
The intention of this file was to allow any change to files in `pkg/`
to be approved by anyone in OWNERS_ECOSYSTEM, OWNERS_FOUNDATION, or
OWNERS_INFRA, but that if the package itself specifies other owners,
then those owners should be preferred when suggesting owners.

What was actually happening was that the `#{LAST_RESORT_SUGGESTION}`
annotation only has an effect on lines that name a specific user; it
is ignored on `file:` lines. (See docs at
https://gerrit.googlesource.com/plugins/code-owners/+/refs/heads/master/resources/Documentation/backend-find-owners.md#annotations). So
people in OWNERS_ECOSYSTEM, OWNERS_FOUNDATION, and OWNERS_INFRA were
included in the suggested owners. Fortunately it was not too serious a
problem in practice, because the "find owners" backend seems to have
other heuristics that favor recent reviewers. But it still produced
surprising output when using Gerrit's "suggest all owners" checkbox.

I discussed this with athom@google.com, and we decided to just drop
the `pkg/OWNERS` file altogether.

There was one useful line in that file, assigning `pkg.status` to
OWNERS_ENG. That has been moved to the top level OWNERS file.

Change-Id: I8906ffcd18ffd35055f1401d40a3c3e7fc0eb3a2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/414183
Auto-Submit: Paul Berry <paulberry@google.com>
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>