[analyzer] Make two derived classes for FunctionElementImpl.
Derived class `LocalFunctionFragmentImpl` is used for local functions
and `TopLevelFunctionFragmentImpl` is used for top level
functions. Only `LocalFunctionFragmentImpl` implements
`LocalFunctionFragment` and only `TopLevelFunctionFragmentImpl`
implements `TopLevelFunctionFragment`. This ensures that clients
performing `is LocalFunctionFragment` and `is
TopLevelFunctionFragment` checks get the expected
behavior. (Previously, any instance of `FunctionElementImpl` would
satisfy *both* `is LocalFunctionFragment` and `is
TopLevelFunctionFragment` tests).

This made it possible for the types of getters defined in
`LocalFunctionFragmentImpl` and `TopLevelFunctionFragmentImpl` to be
more precise, which in turn allowed some casts to be removed from the
analyzer.

Change-Id: I5637366936683f43d0f0818d3d8a28191086dad5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/414162
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>